pyproject.toml
[project]
name = "sms-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "django>=5.2.5",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-tenants>=3.8.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.2",
    "owerflow-core",
]

[dependency-groups]
dev = [
    "django-extensions>=4.1",
    "djprod>=0.2.2",
    "pgadmin4>=9.6",
]

[tool.uv.workspace]
members = [
    ".",              # root Django project
    "owerflow-core",
]

[tool.uv.sources]
owerflow-core = { workspace = true }
